node/node: Move options to options.go
